requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
typing-extensions>=4.8.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# WebSocket support
websockets==12.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0

# CORS
fastapi-cors==0.0.6

# Rate limiting
slowapi==0.1.9

# File handling
python-magic==0.4.27
Pillow==10.1.0

# Environment
python-decouple==3.8

# Optional: For production deployment
gunicorn==21.2.0

